+++ modules/core/src/model/checkbox/dynamic-checkbox.model.ts
@@ -6,7 +6,6 @@ export const DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX = "CHECKBOX";
 export class DynamicCheckboxModel extends DynamicFormControlModel<boolean> {
 
     align: string;
-    //checked: boolean; // actually makes no sense since type of value is boolean
     indeterminate: boolean;
 
     constructor(config: {} = {}, cls?: {}) {
@@ -14,8 +13,11 @@ export class DynamicCheckboxModel extends DynamicFormControlModel<boolean> {
         super(config, cls);
 
         this.align = getValue(config, "align", "start");
-        //this.checked = getValue(config, "checked", false);
         this.indeterminate = getValue(config, "indeterminate", false);
         this.type = DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX;
+        
+        if (this.value !== true) {
+            this.value = false;
+        }
     }
 }